chore(Auth)[OK-57]: Rename files and directories for consistency
