+++ frontend/src/app/MessageReceiver.tsx
@@ -9,7 +9,7 @@
                 //
 
 import React, { useEffect, useRef, useState } from "react";
-import { useAuth } from "@/context/AuthContext";
+import { useAuth } from "@/contexts/AuthContext";
 
 const MessageReceiver: React.FC<{ children: React.ReactNode }> = ({ children }) => {
     const { getAuthMessage, clearAuthMessage } = useAuth();
