+++ frontend/src/app/auth/signup/SignupCard.tsx
@@ -12,7 +12,6 @@ import {
 import GoogleIcon from "@/components/non-lucid-icons/GoogleIcon";
 import { Mail } from "lucide-react";
 import {googleAuth} from "@/lib/googleAuth";
-import GoogleLogin from "@/components/GoogleLogin/GoogleLogin";
 
 const SignupCard: React.FC = () => {
     return (
@@ -35,7 +34,6 @@ const SignupCard: React.FC = () => {
                 <Button variant="outline" className="flex-1" onClick={googleAuth}>
                     <GoogleIcon width={16} height={16} />&ensp;Google
                 </Button>
-                <GoogleLogin/>
             </CardContent>
         </Card>
     );
