+++ frontend/src/app/auth/signup/email/EmailSignupCard.tsx
@@ -3,7 +3,7 @@ import {Button} from "@/components/ui/button";
 import {Input} from "@/components/ui/input"
 import {Label} from "@/components/ui/label"
 import {Card, CardContent, CardDescription, CardHeader, CardTitle,} from "@/components/ui/card"
-import {useHandleAuth} from "@/hooks/auth-form/useHandleAuth";
+import {useHandleAuth} from "@/hooks/auth/useHandleAuth";
 
 const EmailSignupCard: React.FC = () => {
     const { email, password, error, loading, setEmail, setPassword, signup } = useHandleAuth();
